fs working, persistence to-do

+++ src/inode/inode.h
@@ -19,11 +19,11 @@ typedef struct inode_struct
     uint16_t permissions{};
 
     // time of last access
-    uint32_t access_time{};
+    uint64_t access_time{};
     // time of last change (perms or content)
-    uint32_t change_time{};
+    uint64_t change_time{};
     // time of last modification (only content)
-    uint32_t modify_time{};
+    uint64_t modify_time{};
 
     uint32_t links_count{};
